Вынест bokkmarks в отдельный редюсер и сервис. Заменил show/hid в dropdown на toggle.
