+++ src/blocks/account-profile/account-profile.js
@@ -7,7 +7,7 @@ import ProfileHeader from './header/account-profile-header';
 import ProfileFeed from './feed/account-profile-feed';
 import { UserInfo, Tabs } from './../../blocks';
 
-import { myCollectionsLoader, savedLinksLoader } from './../../reducers/user.reducer';
+import { myCollectionsLoader, savedLinksLoader } from './../../reducers/bookmarks.reducer';
 
 import './account-profile.scss';
 
@@ -29,7 +29,7 @@ class AccountProfile extends Component {
     }
 
     render() {
-        const { user, archive, typeToFeed } = this.props;
+        const { user, bookmarks } = this.props;
         const tabs = [
             {
                 id: 1,
@@ -66,22 +66,21 @@ class AccountProfile extends Component {
                 name: 'savedCollections',
             },
         ];
-        const data = typeToFeed.toLowerCase().indexOf('links') > -1 ? archive.links : archive.collections;
+        const data = bookmarks.typeToFeed.toLowerCase().indexOf('links') > -1 ? bookmarks.links : bookmarks.collections;
         return (<div className="account-profile-wrap">
             <ProfileHeader />
             <UserInfo user={user} />
             <div className="account-profile__tabs-wrap">
                 <Tabs tabs={tabs} />
             </div>
-            <ProfileFeed data={data} type={typeToFeed} filterItems={filterItems} />
+            <ProfileFeed data={data} type={bookmarks.typeToFeed} filterItems={filterItems} />
         </div>);
     }
 }
 
 AccountProfile.propTypes = {
     user: PropTypes.object.isRequired,
-    archive: PropTypes.object.isRequired,
-    typeToFeed: PropTypes.string.isRequired,
+    bookmarks: PropTypes.object.isRequired,
     token: PropTypes.string.isRequired,
     myCollectionsLoader: PropTypes.func.isRequired,
     savedLinksLoader: PropTypes.func.isRequired,
@@ -92,7 +91,7 @@ function mapStateToProps(state) {
     return {
         user: state.user.data,
         token: state.app.token,
-        archive: state.user.archive,
+        bookmarks: state.bookmarks,
         typeToFeed: state.user.typeToFeed,
     };
 }
