+++ src/reducers/index.js
@@ -8,6 +8,7 @@ import { reducer as createCollection } from './create-collection.reducer';
 import { reducer as modal } from './modal.reducer';
 import { reducer as link } from './link.reducer';
 import { reducer as loader } from './loader.reducer';
+import { reducer as bookmarks } from './bookmarks.reducer';
 
 export default combineReducers({
     feed,
@@ -19,4 +20,5 @@ export default combineReducers({
     modal,
     link,
     loader,
+    bookmarks,
 });
